+++ packages/project/src/Workspace.ts
@@ -13,11 +13,11 @@ const PROJECT_EXTENSIONS = ['.yaml', '.yml'];
 export class Workspace {
   private config?: OpenfnConfig;
   private projects: Project[] = [];
+  private projectPaths = new Map<string, string>();
   private isValid: boolean = false;
   constructor(workspacePath: string) {
     const projectsPath = path.join(workspacePath, PROJECTS_DIRECTORY);
     const openfnYamlPath = path.join(workspacePath, OPENFN_YAML_FILE);
-
     // dealing with openfn.yaml
     if (pathExists(openfnYamlPath, 'file')) {
       this.isValid = true;
@@ -34,8 +34,11 @@ export class Workspace {
         );
 
       this.projects = stateFiles.map((file) => {
-        const data = fs.readFileSync(path.join(projectsPath, file), 'utf-8');
-        return fromAppState(data, { format: 'yaml' });
+        const stateFilePath = path.join(projectsPath, file);
+        const data = fs.readFileSync(stateFilePath, 'utf-8');
+        const project = fromAppState(data, { format: 'yaml' });
+        this.projectPaths.set(project.name, stateFilePath);
+        return project;
       });
     }
   }
@@ -46,6 +49,10 @@ export class Workspace {
 
   get(id: string) {
     return this.projects.find((p) => p.name === id);
+  }
+
+  getProjectPath(id: string) {
+    return this.projectPaths.get(id);
   }
 
   getActiveProject() {